Clipboard and Settings
Clipboard

When a row is clicked the password contained in the selected row is
copied to the clipboard. On X11 it is also copied to the mouse selection.

Settings

The location of the db file is configurable now. If no configuration
is found the user is asked where to store it. This setting is then
saved using `QSettings`. On unix this is stored in
`~/.config/<orgName>/`.